crypto-tracker-bot/requirements.txt
supabase==2.7.1
python-telegram-bot==20.3
web3==6.15.1
solana==0.29.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
aioredis==2.0.1
pytz==2024.1
python-socketio==5.11.2
eth-account==0.11.0
trx-utils==0.2.2
eospy==0.8.1
near-api-py==0.8.0
algorand-sdk==2.4.0
ton==0.43.1
polkadot==0.4.2
dogecoin==0.1.2
cosmos-sdk==0.2.0
osmosis-sdk==0.1.0